refactor: migrate automated tasks to ts
