+++ .github/workflows/issuesAutomatedTasks.yml
@@ -162,7 +162,7 @@ jobs:
               sudo rm -rf "$FOLDER_PATH"
           - name: 💿 Do actions in database
             run: |
-                node scripts/automatedTasks.mjs "$OPERATION" "$REQUEST"
+                tsx scripts/automatedTasks.mjs "$OPERATION" "$REQUEST"
           - name: 🤖 Update JSON files after this change
             run: npm run generate-api-json-files
           - name: 🐛 Fix corner cases to make creation of pull request possible
