+++ package.json
@@ -46,9 +46,11 @@
   },
   "devDependencies": {
     "@inquirer/input": "^5.1.2",
+    "@types/better-sqlite3": "^7.6.13",
     "better-sqlite3": "^12.6.2",
     "feed": "^5.2.1",
-    "googleapis": "^173.0.0"
+    "googleapis": "^173.0.0",
+    "tsx": "^4.22.4"
   },
   "overrides": {
     "@types/react": "19.2.17",
