+++ scripts/tasks/delete-backlog.ts
@@ -1,0 +1,10 @@
+import type { Database } from 'better-sqlite3';
+
+type DeleteBacklogPayload = {
+  title: string;
+};
+
+export async function deleteBacklogFromDatabase(db: Database, payload: DeleteBacklogPayload) {
+  const deleteBacklogStmt = db.prepare("DELETE FROM backlog WHERE title = ?");
+  return deleteBacklogStmt.run(payload.title);
+}
