+++ scripts/tasks/index.ts
@@ -1,0 +1,12 @@
+export * from './add-game';
+export * from './update-game';
+export * from './delete-game';
+export * from './add-backlog';
+export * from './delete-backlog';
+export * from './clean-backlog';
+export * from './add-serie';
+export * from './manage-serie';
+export * from './manage-dlcs';
+export * from './add-test';
+export * from './update-test';
+export * from './delete-test';
